firstlook/scripts/requirements.txt
paho-mqtt>=1.6.1
cbor2>=5.4.6
